WebApi | added Delete control  relational datas

+++ WebApi/Application/ActorOperations/Commands/DeleteActor/DeleteActorCommand.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using Microsoft.EntityFrameworkCore;
 using WebApi.DbOperations;
 using WebApi.Entities;
 
@@ -19,11 +20,14 @@ public class DeleteActorCommand
 
     public void Handle()
     {
-        var actorInDb = context.Actors.SingleOrDefault(m=>m.Id == ActorId);
+        var actorInDb = context.Actors.Include(a=> a.Movies).SingleOrDefault(m=>m.Id == ActorId);
 
         if(actorInDb is null)
             throw new InvalidOperationException("ActorId: "+ActorId+" does not exists.");
         
+        if(actorInDb.Movies.Any())
+            throw new InvalidOperationException("ActorId: " + ActorId + " has " +actorInDb.Movies.Count()+ " movies. Please delete them first.");
+            
         context.Actors.Remove(actorInDb);
         
         context.SaveChanges();
